% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeacoQC.R
\name{PlotPeacoQC}
\alias{PlotPeacoQC}
\title{Visualise deleted cells of PeacoQC}
\usage{
PlotPeacoQC(ff, channels, output_directory = ".", display_cells = 5000,
            manual_cells = NULL, title_FR = NULL, peaks = TRUE,
            prefix = "PeacoQC_")
}
\arguments{
\item{ff}{A flowframe}

\item{channels}{Indices of the channels in the ff that have to be plotted}

\item{output_directory}{Directory where the plots should be generated. Set to NULL if no plots need to be generated}

\item{display_cells}{The number of measurements that should be displayed. (The number of dots that are displayed for every channel)}

\item{manual_cells}{THIS HAS TO BE REMOVED FOR THE FINAL VERSION}

\item{title_FR}{The title that has to be displayed above the flow rate figure}

\item{peaks}{If set to TRUE: \code{PeacoQC} will be run and the peaks will be displayed. If the result of \code{PeacoQC} is given, all the quality control results will be visualised.}

\item{prefix}{The prefix that will be given to the generated png file}
}
\value{
This function returns nothing but generates a png file in the output_directory
}
\description{
\code{PlotPeacoQC} will generate a png file with on overview of the flow rate and the different selected channels. These will be annotated based on the measurements that were removed by PeacoQC. It is also possible to only display the quantiles and median or only the measurements without any annotation.
}
\examples{
\dontrun{
# Read in flowframe SHOULD BE  ADAPTED!!!!!!!!!!!
ff <- read.FCS(file)
# Determine the channels on which quality control should be performed
channels <- colnames(ff@exprs)[c(1,3,5:10)]
peacoqc_res <- PeacoQC(ff = ff, channels = channels, determine_good_cells = TRUE,
                       output_directory = ".", plot = TRUE)
}
}
