% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeacoQC.R
\name{RemoveMargins}
\alias{RemoveMargins}
\title{Remove margin events}
\usage{
RemoveMargins(ff, channels, channel_specifications = NULL, output = "frame")
}
\arguments{
\item{ff}{A flowframe}

\item{channels}{The channels that have to be checked for margin events}

\item{channel_specifications}{A list of lists with parameter specifications per channel. This parameter should only be used if the values in the internal parameters description is too strict. There should be one list per channel with first a minRange and then a maxRange value. This list should have the channel name.}

\item{output}{If set to "full", a list with the filtered flowframe and the indices of the margin event is returned. If set to "frame", only the filtered flowframe is returned.}
}
\description{
\code{RemoveMargins} will remove margin events from the flowframe based on the internal description of the fcs file.
}
