% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeacoQC.R
\name{PeacoQCHeatmap}
\alias{PeacoQCHeatmap}
\title{Make overview heatmap of quality control analysis}
\usage{
PeacoQCHeatmap(report_location, heatmap_name = "PeacoQC_heatmap",
display_plot = FALSE, output_directory = ".",
latest_tests = FALSE, title = "PeacoQC report")
}
\arguments{
\item{report_location}{The path to the PeacoQC report generated by \code{PeacoQC}.}

\item{heatmap_name}{The name that png file will have with the heatmap.}

\item{display_plot}{If set to TRUE, the heatmap will be displayed in the R environment.}

\item{output_directory}{Directory where the heatmap will be generated. If set to NULL, the heatmap will not be saved as png file.}

\item{latest_tests}{If this is set to TRUE, only the latest quality control run will be displayed in the heatmap.}

\item{title}{The title that should be given to the heatmap.}
}
\value{
This function returns nothing but generates a png file in the output_directory with the heatmap.
}
\description{
\code{PeacoQCHeatmap} will make a heatmap to display all the results generated by \code{PeacoQC}. It will include the percentages of measurements that are removed in total, by the IT method and by the MAD method. It will also show the parameters that were used during the quality control.
}
\examples{
\dontrun{
SHOULD BE ADJUSTED
}
}
